+++ apps/gateway/src/modules/career/crafts/stores/stores.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+
+import { StoresResolver } from './stores.resolver';
+import { StoresController } from './stores.controller';
+
+@Module({
+  controllers: [StoresController],
+  providers: [StoresResolver],
+})
+export class StoresModule {}